Introduce a new help system
This is the basic structure for a help system that allows context aware
help buttons to be shown throughout the site. Each button is configured
separately but activates the shared help modal. The content of the modal
is configured for each help button which makes it possible to show only
the help content relevant to the specific button placement.